+++ xvenv/const.py
@@ -1,1 +1,1 @@
-VERSION = "v0.0.3-rc1"
+VERSION = "v0.0.3"