+++ CHANGELOG.md
@@ -5,11 +5,13 @@ for open development tasks and limitations.
 
 # CHANGELOG
 
+
 ## version v0.0.3 - ???
 
 - `-venv` cmd-line parameter in addition to `-cwd` parameter
   - `-venv` specifies the venv folder, and 
   `-cwd` is working directory folder
+- pip update after ensurepip added
 -